refactor: simplify and flatten package structure
