+++ commerce/shopping-cart/src/main/java/ru/yandex/practicum/smarthometech/commerce/cart/domain/CartStatus.java
@@ -1,0 +1,6 @@
+package ru.yandex.practicum.smarthometech.commerce.cart.domain;
+
+public enum CartStatus {
+    ACTIVE,
+    INACTIVE
+}
