+++ commerce/shopping-cart/src/main/java/ru/yandex/practicum/smarthometech/commerce/cart/domain/ShoppingCartRepository.java
@@ -1,0 +1,14 @@
+package ru.yandex.practicum.smarthometech.commerce.cart.domain;
+
+import java.util.Optional;
+import java.util.UUID;
+
+public interface ShoppingCartRepository {
+
+    Optional<ShoppingCart> findActiveByUsername(String username);
+
+    Optional<ShoppingCart> findById(UUID shoppingCartId);
+
+    ShoppingCart save(ShoppingCart cart);
+
+}
