--- commerce/shopping-cart/src/main/java/ru/yandex/practicum/smarthometech/commerce/cart/domain/enums/CartStatus.java
@@ -1,6 +1,0 @@
-package ru.yandex.practicum.smarthometech.commerce.cart.domain.enums;
-
-public enum CartStatus {
-    ACTIVE,
-    INACTIVE
-}
