--- commerce/shopping-cart/src/main/java/ru/yandex/practicum/smarthometech/commerce/cart/domain/repository/ShoppingCartRepository.java
@@ -1,16 +1,0 @@
-package ru.yandex.practicum.smarthometech.commerce.cart.domain.repository;
-
-import ru.yandex.practicum.smarthometech.commerce.cart.domain.entity.ShoppingCart;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface ShoppingCartRepository {
-
-    Optional<ShoppingCart> findActiveByUsername(String username);
-
-    Optional<ShoppingCart> findById(UUID shoppingCartId);
-
-    ShoppingCart save(ShoppingCart cart);
-
-}
