--- commerce/shopping-store/src/main/java/ru/yandex/practicum/smarthometech/commerce/store/application/exception/ProductNotFoundException.java
@@ -1,19 +1,0 @@
-package ru.yandex.practicum.smarthometech.commerce.store.application.exception;
-
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-import java.util.UUID;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-@Getter
-public class ProductNotFoundException extends RuntimeException {
-
-    private final UUID productId;
-
-    public ProductNotFoundException(UUID productId) {
-        super(String.format("Product with ID '%s' not found.", productId));
-        this.productId = productId;
-    }
-}
