+++ commerce/shopping-store/src/main/java/ru/yandex/practicum/smarthometech/commerce/store/domain/ProductCategory.java
@@ -1,0 +1,7 @@
+package ru.yandex.practicum.smarthometech.commerce.store.domain;
+
+public enum ProductCategory {
+    LIGHTING,
+    CONTROL,
+    SENSORS
+}
