+++ commerce/shopping-store/src/main/java/ru/yandex/practicum/smarthometech/commerce/store/domain/ProductState.java
@@ -1,0 +1,6 @@
+package ru.yandex.practicum.smarthometech.commerce.store.domain;
+
+public enum ProductState {
+    ACTIVE,
+    DEACTIVATE
+}
