--- commerce/shopping-store/src/main/java/ru/yandex/practicum/smarthometech/commerce/store/domain/enums/ProductCategory.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.smarthometech.commerce.store.domain.enums;
-
-public enum ProductCategory {
-    LIGHTING,
-    CONTROL,
-    SENSORS
-}
