--- commerce/shopping-store/src/main/java/ru/yandex/practicum/smarthometech/commerce/store/domain/enums/ProductState.java
@@ -1,6 +1,0 @@
-package ru.yandex.practicum.smarthometech.commerce.store.domain.enums;
-
-public enum ProductState {
-    ACTIVE,
-    DEACTIVATE
-}
