--- commerce/shopping-store/src/main/java/ru/yandex/practicum/smarthometech/commerce/store/domain/enums/QuantityState.java
@@ -1,8 +1,0 @@
-package ru.yandex.practicum.smarthometech.commerce.store.domain.enums;
-
-public enum QuantityState {
-    ENDED,
-    FEW,
-    ENOUGH,
-    MANY
-}
