--- commerce/shopping-store/src/main/java/ru/yandex/practicum/smarthometech/commerce/store/domain/repository/ProductRepository.java
@@ -1,18 +1,0 @@
-package ru.yandex.practicum.smarthometech.commerce.store.domain.repository;
-
-import java.util.Optional;
-import java.util.UUID;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import ru.yandex.practicum.smarthometech.commerce.store.domain.enums.ProductCategory;
-import ru.yandex.practicum.smarthometech.commerce.store.domain.enums.ProductState;
-import ru.yandex.practicum.smarthometech.commerce.store.domain.entity.Product;
-
-public interface ProductRepository {
-
-    Optional<Product> findById(UUID productId);
-    Product save(Product product);
-    Page<Product> findByCategoryAndState(ProductCategory category, ProductState state, Pageable pageable);
-    boolean existsById(UUID productId);
-
-}
