+++ commerce/shopping-store/src/main/java/ru/yandex/practicum/smarthometech/commerce/store/presentation/ShoppingStoreController.java
@@ -1,0 +1,63 @@
+package ru.yandex.practicum.smarthometech.commerce.store.presentation;
+
+import java.util.UUID;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import ru.yandex.practicum.smarthometech.commerce.api.client.ShoppingStoreClient;
+import ru.yandex.practicum.smarthometech.commerce.api.dto.store.ProductCategory;
+import ru.yandex.practicum.smarthometech.commerce.api.dto.store.ProductDto;
+import ru.yandex.practicum.smarthometech.commerce.api.dto.store.SetProductQuantityStateRequest;
+import ru.yandex.practicum.smarthometech.commerce.store.application.ShoppingStoreService;
+
+@RestController
+@RequestMapping("/api/v1/shopping-store")
+@RequiredArgsConstructor
+public class ShoppingStoreController implements ShoppingStoreClient {
+
+    private final ShoppingStoreService shoppingStoreService;
+
+    @Override
+    @GetMapping
+    public Page<ProductDto> getProducts(@RequestParam ProductCategory category, Pageable pageable) {
+        return shoppingStoreService.getProducts(category, pageable);
+    }
+
+    @Override
+    @PutMapping
+    public ProductDto createNewProduct(@RequestBody ProductDto productDto) {
+        return shoppingStoreService.createNewProduct(productDto);
+    }
+
+    @Override
+    @PostMapping
+    public ProductDto updateProduct(@RequestBody ProductDto productDto) {
+        return shoppingStoreService.updateProduct(productDto);
+    }
+
+    @Override
+    @PostMapping("/removeProductFromStore")
+    public boolean removeProduct(@RequestBody UUID productId) {
+        return shoppingStoreService.removeProductFromStore(productId);
+    }
+
+    @Override
+    @PostMapping("/quantityState")
+    public boolean setStatus(@RequestBody SetProductQuantityStateRequest request) {
+        return shoppingStoreService.setProductQuantityState(request);
+    }
+
+    @Override
+    @GetMapping("/{productId}")
+    public ProductDto getProduct(@PathVariable UUID productId) {
+        return shoppingStoreService.getProduct(productId);
+    }
+}
