+++ commerce/warehouse/src/main/java/ru/yandex/practicum/smarthometech/commerce/warehouse/application/ProductNotFoundException.java
@@ -1,0 +1,19 @@
+package ru.yandex.practicum.smarthometech.commerce.warehouse.application;
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+import java.util.UUID;
+
+@ResponseStatus(value = HttpStatus.NOT_FOUND)
+@Getter
+public class ProductNotFoundException extends RuntimeException {
+
+    private final UUID productId;
+
+    public ProductNotFoundException(UUID productId) {
+        super(String.format("Product with ID '%s' not found in warehouse.", productId));
+        this.productId = productId;
+    }
+}
