--- commerce/warehouse/src/main/java/ru/yandex/practicum/smarthometech/commerce/warehouse/application/mapper/WarehouseMapper.java
@@ -1,25 +1,0 @@
-package ru.yandex.practicum.smarthometech.commerce.warehouse.application.mapper;
-
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-import ru.yandex.practicum.smarthometech.commerce.api.dto.warehouse.DimensionDto;
-import ru.yandex.practicum.smarthometech.commerce.api.dto.warehouse.NewProductInWarehouseRequest;
-import ru.yandex.practicum.smarthometech.commerce.warehouse.domain.entity.WarehouseItem;
-
-@Mapper(componentModel = "spring")
-public interface WarehouseMapper {
-
-    @Mapping(target = "productId", source = "productId")
-    @Mapping(target = "quantity", constant = "0L")
-    @Mapping(target = "weightKg", source = "weight")
-    @Mapping(target = "isFragile", source = "fragile")
-    @Mapping(target = "widthM", source = "dimension.width")
-    @Mapping(target = "heightM", source = "dimension.height")
-    @Mapping(target = "depthM", source = "dimension.depth")
-    WarehouseItem newProductRequestToWarehouseItem(NewProductInWarehouseRequest request);
-
-    @Mapping(target = "width", source = "widthM")
-    @Mapping(target = "height", source = "heightM")
-    @Mapping(target = "depth", source = "depthM")
-    DimensionDto warehouseItemToDimensionDto(WarehouseItem item);
-}
