+++ commerce/warehouse/src/main/java/ru/yandex/practicum/smarthometech/commerce/warehouse/infrastructure/JpaWarehouseRepositoryInterface.java
@@ -1,0 +1,8 @@
+package ru.yandex.practicum.smarthometech.commerce.warehouse.infrastructure;
+
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+import ru.yandex.practicum.smarthometech.commerce.warehouse.domain.WarehouseItem;
+
+public interface JpaWarehouseRepositoryInterface extends JpaRepository<WarehouseItem, UUID> {
+}
