--- commerce/warehouse/src/main/java/ru/yandex/practicum/smarthometech/commerce/warehouse/infrastructure/repository/JpaWarehouseRepositoryAdapter.java
@@ -1,36 +1,0 @@
-package ru.yandex.practicum.smarthometech.commerce.warehouse.infrastructure.repository;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-import ru.yandex.practicum.smarthometech.commerce.warehouse.domain.entity.WarehouseItem;
-import ru.yandex.practicum.smarthometech.commerce.warehouse.domain.repository.WarehouseRepository;
-
-@Repository
-@RequiredArgsConstructor
-public class JpaWarehouseRepositoryAdapter implements WarehouseRepository {
-
-    private final JpaWarehouseRepositoryInterface jpaRepository;
-
-    @Override
-    public Optional<WarehouseItem> findById(UUID productId) {
-        return jpaRepository.findById(productId);
-    }
-
-    @Override
-    public List<WarehouseItem> findAllById(Iterable<UUID> productIds) {
-        return jpaRepository.findAllById(productIds);
-    }
-
-    @Override
-    public WarehouseItem save(WarehouseItem warehouseItem) {
-        return jpaRepository.saveAndFlush(warehouseItem);
-    }
-
-    @Override
-    public boolean existsById(UUID productId) {
-        return jpaRepository.existsById(productId);
-    }
-}
